requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
frozenlist==1.7.0
google-api-core==2.25.0
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
multidict==6.4.4
oauthlib==3.2.2
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
py-cord==2.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dotenv==1.1.0
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
yarl==1.20.1
